Move the hooks
