+++ apps/dashboard/tsconfig.json
@@ -8,7 +8,7 @@
       "@components/*": ["./src/app/components/*"],
       "@data/*": ["./src/client-data/data/*"],
       "@actions/*": ["./src/app/actions/*"],
-      "@hooks": ["./src/app/hooks"]
+      "@hooks": ["./src/hooks"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
